message_be/apps/message/views_container/__init__.py
from drf_yasg import openapi
from drf_yasg.utils import swagger_auto_schema

from rest_framework import (viewsets, mixins, status, permissions, generics)
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.viewsets import GenericViewSet
from rest_framework.permissions import IsAuthenticated
from rest_framework.pagination import LimitOffsetPagination
from rest_framework.generics import GenericAPIView, RetrieveAPIView, ListAPIView, CreateAPIView, UpdateAPIView

from message_be.apps.message.models import *
