pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Datbe2001 <datbe26092001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-rest-swagger = "^2.2.0"
django = "4.2.7"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
mongoengine = "^0.28.2"
dj-rest-auth = "6.0.0"
channels = "^4.1.0"
channels-redis = "^4.2.0"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
